backend/requirement.txt
Flask>=2.2.2,<2.3
werkzeug<2.3
flask-cors==3.0.10
torch>=2.0.0
torchaudio>=2.0.0
soundfile
numpy
scipy
inflect
librosa
progressbar2
rotary-embedding-torch
unidecode
TTS
llama-cpp-python==0.2.56
git+https://github.com/suno-ai/bark.git
nltk